pyproject.toml
[project]
name = "trec-product"
version = "0.1.0"
description = "Tools for the TREC product search & recommendation track."
readme = "README.md"
requires-python = "~=3.12.0"
dependencies = [
  "docopt-ng ~=0.9",
  "duckdb ~=1.3",
  "humanize>=4.12.3",
  "pandas>=2.2.3",
  "xopen>=2.0.2",
]
